+++ accessories/currentConditions.js
@@ -1,5 +1,6 @@
 const debug = require('debug')('homebridge-weather-plus'),
-	version = require("../package.json").version;
+	version = require("../package.json").version,
+	compatibility = require("../util/compatibility");
 
 let Service,
 	Characteristic,
@@ -26,7 +27,6 @@ function CurrentConditionsWeatherAccessory(platform, stationIndex)
 	this.name = this.config.nameNow;
 	this.displayName = this.config.nameNow;
 	this.stationIndex = stationIndex;
-	this.compatibilityTypes = this.config.compatibilityTypes;
 
 	// Use homekit temperature service or eve weather service depending on compatibility setting
 	debug("Using compatibility mode '%s'", this.config.compatibility);
@@ -37,17 +37,7 @@ function CurrentConditionsWeatherAccessory(platform, stationIndex)
 	else if (this.config.compatibility === "home")
 	{
 		this.CurrentConditionsService = new Service.TemperatureSensor(this.name + " Temperature", "Temperature");
-
-		this.AirPressureService = new Service.OccupancySensor("Air Pressure", "Air Pressure");
-		this.CloudCoverService = new Service.OccupancySensor("Cloud Cover", "Cloud Cover");
-		this.DewPointService = new Service.TemperatureSensor("Dew Point", "Dew Point");
-		this.HumidityService = new Service.HumiditySensor("Humidity");
-		this.RainBoolService = new Service.OccupancySensor("Rain", "Rain");
-		this.SnowBoolService = new Service.OccupancySensor("Snow", "Snow");
-		this.UVIndexService = new Service.AirQualitySensor("UV Index", "UV Index");
-		this.VisibilityService = new Service.OccupancySensor("Visibility", "Visibility");
-		this.WindDirectionService = new Service.OccupancySensor("Wind Direction", "Wind Dir");
-		this.WindSpeedService = new Service.OccupancySensor("Wind Speed", "Wind Speed");
+		compatibility.createServices(this, Service);
 	}
 	else
 	{
@@ -74,30 +64,16 @@ function CurrentConditionsWeatherAccessory(platform, stationIndex)
 				this.CurrentConditionsService.getCharacteristic(Characteristic.CurrentTemperature).props.minValue = -50;
 			}
 			// Use separate services for these characteristics if compatiblity is "home"
-			else if (this.config.compatibility === "home" && this.compatibilityTypes.includes(name))
+			else if (this.config.compatibility === "home" && compatibility.types.includes(name))
 			{
-				// Workaround to get configured units
-				if (name === "Visibility")
-				{
-					this.VisibilityService.addCharacteristic(CustomCharacteristic.Visibility);
-					let unit = this.VisibilityService.getCharacteristic(CustomCharacteristic.Visibility).props.unit;
-					this.VisibilityService.removeCharacteristic(CustomCharacteristic.Visibility);
-					this.VisibilityService.unit = unit;
-				}
-				else if (name === "WindSpeed")
-				{
-					this.WindSpeedService.addCharacteristic(CustomCharacteristic.WindSpeed);
-					let unit = this.WindSpeedService.getCharacteristic(CustomCharacteristic.WindSpeed).props.unit;
-					this.WindSpeedService.removeCharacteristic(CustomCharacteristic.WindSpeed);
-					this.WindSpeedService.unit = unit;
-				}
+				compatibility.customizeServices(this, CustomCharacteristic, name);
 			}
-			// Use separate service for this characteristic if configured
+			// Use separate service for humidity if configured
 			else if (this.config.compatibility === "eve" && name === "Humidity" && this.config.extraHumidity)
 			{
 				//
 			}
-			// Use separate service for this characteristic if configured
+			// Add humidity characteristic to temperature service if configured
 			else if (name === "Humidity")
 			{
 				this.CurrentConditionsService.addCharacteristic(Characteristic.CurrentRelativeHumidity);
@@ -130,26 +106,6 @@ CurrentConditionsWeatherAccessory.prototype = {
 
 	getServices: function ()
 	{
-		let services = [this.informationService, this.CurrentConditionsService, this.historyService];
-
-		// for (let i = 0; i < customServices.length; i++)
-		// {
-		// 	let service = customServices[i] + "Service";
-		// 	if (service in this)
-		// 	{
-		// 		services.push(this.service);
-		// 	}
-		// }
-
-		this.compatibilityTypes.forEach((name) =>
-		{
-			let service = name + "Service";
-			if (service in this)
-			{
-				services.push(this[service]);
-			}
-		});
-
-		return services;
+		return [this.informationService, this.CurrentConditionsService, this.historyService].concat(compatibility.getServices(this));
 	}
 };